refactor(model): 包化共享音频实体入口
Why:
- 让 `model` 目录语义先与当前唯一成型的共享领域对象对齐
- 为后续独立收口 task helper 或进一步模型化预留稳定目录壳

What:
- 将 `AudioEntityData` 迁入 `model/entity.py`，并把 `model.py` 改为 `model/` 目录包
- 在 `model/__init__.py` 保留 `AudioEntityData` 与 `generate_*_tasks` 导出，减少现有调用点改动
- 新增 `tests/test_model_package_api.py` 作为 `model` 目录包公开导出护栏

Risk:
- 兼容性影响：无，对外仍保持 `lol_audio_unpack.model` 导入入口
- 潜在风险：`generate_*_tasks` 仍暂留在 `__init__.py`，后续若继续拆分需再补独立护栏

Test:
- `uv run ruff check src/lol_audio_unpack/model tests/test_model_package_api.py tests/test_context_injection.py tests/test_unpack_bp_vo.py tests/test_champion_visibility.py tests/test_mapping.py`
- `uv run pytest -q tests/test_model_package_api.py tests/test_context_injection.py tests/test_unpack_bp_vo.py tests/test_champion_visibility.py tests/test_mapping.py`
